+++ NEWS.rst
@@ -1,3 +1,46 @@
+v8.7.1
+======
+
+Bugfixes
+--------
+
+- Fixed errors in FastPath under fork-multiprocessing. (#520)
+- Removed cruft from Python 3.8. (#524)
+
+
+v8.7.0
+======
+
+Features
+--------
+
+- ``.metadata()`` (and ``Distribution.metadata``) can now return ``None`` if the metadata directory exists but not metadata file is present. (#493)
+
+
+Bugfixes
+--------
+
+- Raise consistent ValueError for invalid EntryPoint.value (#518)
+
+
+v8.6.1
+======
+
+Bugfixes
+--------
+
+- Fixed indentation logic to also honor blank lines.
+
+
+v8.6.0
+======
+
+Features
+--------
+
+- Add support for rendering metadata where some fields have newlines (python/cpython#119650).
+
+
 v8.5.0
 ======